Update packaging for test
