+++ src/test/java/Flight/FlightManagerTest.java
@@ -1,3 +1,5 @@
+package Flight;
+
 import Flight.FlightManager;
 import org.junit.Before;
 import org.junit.Test;
