+++ src/test/java/Ticket/TicketTest.java
@@ -1,3 +1,5 @@
+package Ticket;
+
 import Ticket.Ticket;
 import org.junit.*;
 import java.time.LocalDateTime;
